requirements.txt
websockets==10.0
PyJWT==2.3.0
